+++ SlimHUD/Controllers/SettingsController.swift
