+++ SlimHUD/Errors/Errors.swift
@@ -1,0 +1,13 @@
+//
+//  Errors.swift
+//  SlimHUD
+//
+//  Created by Alex Perathoner on 19/12/2022.
+//  Copyright © 2022 Alex Perathoner. All rights reserved.
+//
+
+import Foundation
+
+struct ParameterError: Error {
+    let message: String
+}
