+++ SlimHUD/Extensions/Extensions.swift
@@ -1,0 +1,31 @@
+//
+//  Extensions.swift
+//  SlimHUD
+//
+//  Created by Alex Perathoner on 19/03/2020.
+//  Copyright © 2020 Alex Perathoner. All rights reserved.
+//
+
+import Cocoa
+import AppKit
+
+
+
+//extension NSImage {
+//	/// https://stackoverflow.com/a/50074538/6884062
+//	/// - Returns: returns the tinted version of a template image
+//	func tint(color: NSColor) -> NSImage {
+//		let image = self.copy() as! NSImage
+//		image.lockFocus()
+//
+//		color.set()
+//
+//		let imageRect = NSRect(origin: NSZeroPoint, size: image.size)
+//		imageRect.fill(using: .sourceAtop)
+//
+//		image.unlockFocus()
+//
+//		return image
+//	}
+//}
+
