+++ SlimHUD/Services/AppleScript.swift
@@ -15,23 +15,6 @@ func runAS(script: String) -> String? {
 	return nil
 }
 
-func isMuted() -> Bool {
-	return runAS(script: "return output muted of (get volume settings)") == "true"
-}
-
-func getVolumeSettings() -> (outPutVolume: Int, inputVolume: Int, alertVolume: Int, outputMuted: Bool) {
-	let o1 = Int(runAS(script: "return output volume of (get volume settings)") ?? "1")!
-	let o2 = Int(runAS(script: "return input volume of (get volume settings)") ?? "1")!
-	let o3 = Int(runAS(script: "return alert volume of (get volume settings)") ?? "1")!
-	let o4 = runAS(script: "return output muted of (get volume settings)") == "true"
-
-	return (o1, o2, o3, o4)
-}
-
-func getOutputVolume() -> Float {
-	return Float(runAS(script: "return output volume of (get volume settings)") ?? "1")! / 100.0
-}
-
 
 func getDisplayBrightness() -> Float {
 	var brightness: float_t = 1
