+++ SlimHUD/Services/LoginItemList.swift
