+++ SlimHUD/Views/About/AboutViewController.swift
