+++ SlimHUD/Views/About/AboutWindowController.swift
