+++ SlimHUD/Views/BarView/BarView.xib
