+++ SlimHUD/Views/Hud.swift
