+++ SlimHUD/Views/Settings/SettingsViewController/SettingsViewController+General.swift
