+++ SlimHUD/Views/Settings/SettingsViewController/SettingsViewController+Position.swift
