+++ SlimHUD/Views/Settings/SettingsViewController/SettingsViewController+Style.swift
