+++ SlimHUD/Views/Settings/SettingsViewController/SettingsViewController.swift
