+++ SlimHUD/Views/Settings/SettingsWindow.xib
