+++ SlimHUD/Views/Settings/SettingsWindowController.swift
