+++ SlimHUD/Views/Settings/SettingsWindowControllerDelegate.swift
